requirements.txt
docopt==0.6.2
pycairo==1.23.0
pytest==7.3.1
tqdm==4.65.0
